requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
gpt4all
pydantic
PyPDF2